一覧画面の作成

+++ bbs/index.php
@@ -10,10 +10,11 @@ if(isset($_SESSION['id']) && isset($_SESSION['name'])) {
     exit();
 }
 
+$db = dbConnect();
+
 //メッセージの投稿
 if($_SERVER['REQUEST_METHOD'] === 'POST') {
     $message = filter_input(INPUT_POST, 'message', FILTER_SANITIZE_STRING);
-    $db = dbConnect();
     $stmt =  $db->prepare('insert into posts (message, member_id) values(?, ?)');
     if(!$stmt) {
         die($db->error);
@@ -61,13 +62,28 @@ if($_SERVER['REQUEST_METHOD'] === 'POST') {
             </div>
         </form>
 
+        <?php $stmt = $db->prepare('select p.id, p.member_id, p.message, p.created, m.name, m.picture from posts p, members m where m.id=p.member_id order by id desc');
+        if(!$stmt) {
+            die($db->error);
+        }
+        $success = $stmt->execute();
+        if(!$success) {
+            die($db->error);
+        }
+
+        $stmt->bind_result($id, $member_id, $message, $created, $name, $picture);
+        while($stmt->fetch()):
+        ?>
         <div class="msg">
-            <img src="member_picture/" width="48" height="48" alt=""/>
-            <p>○○<span class="name">（○○）</span></p>
-            <p class="day"><a href="view.php?id=">2021/01/01 00:00:00</a>
+            <?php if($picture): ?>
+                <img src="member_picture/<?php echo hsc($picture); ?>" width="48" height="48" alt=""/>
+            <?php endif; ?>
+            <p><?php echo hsc($message); ?><span class="name">（<?php echo hsc($name); ?>）</span></p>
+            <p class="day"><a href="view.php?id="><?php echo hsc($created); ?></a>
                 [<a href="delete.php?id=" style="color: #F33;">削除</a>]
             </p>
         </div>
+        <?php endwhile; ?>
     </div>
 </div>
 </body>
